restructures routes and controllers
